Add files via upload
The designs are quite distinct from each other. 2 are more of the OLS brand image and the other is just me trying to risk a different approach. I would like feedback on the most preferred and I can share the svgs for changes to be made because I'm still figuring open source out, thanks.