Add files via upload
This PR is referencing Issue #17

While making research and as stated in the issue description, Kenaz is a symbol in rune that describes many things. 
I took to google (The most knowing, lol) and there I found out that Kenaz has a symbol on its own. Using the symbol as a base for the first draft, I made a design.

It is in such a way that anyone who has the knowledge about the rune can identify, the human figures symbolises the coming together of the people from different educational background to give back to the world. 
Included here are PNG format for viewing and mock ups of what they look like on merch.

For the second logo, the goal was to immitate the Kenaz rune design without using the actual rune design. I can call it a work in progress though. The bulb is a representation of knowledge ( one of the meaning of the Kenaz symbol) to symbolize the knowledge cohorts are giving back to the world.

Overall, I would say personally that I prefer the first logo (my bias). It was also kowlegeable working on this as this is my first time working to actively create a logo.