Add files via upload
Social Preview Image for Github open-life-science#4
After researching what a social preview is, I got to understand that social preview are cards that give a preview of what a shared link will lead to across all social media page like Twitter, slack, and linked-in, etc., this is also eye catchy which make users get interested to click and see what that sites actually look like.
Well, I just got into UI/UX design a few months ago and for now, this design was made using the Figma App which I am more familiar with and also comfortable with for now, and I also hope to explore other Apps for better usage and understanding as time goes on, The colors used for this design are Green with a color code of #28a82d, Blue with a code of #287aab, White and Black with an opacity of 100% and 95% respectively, Using the Montserrat font for all texts using 36px for the header and 24px for body text. However, I tried to check the project guides on Colors, Fonts, and other tools but couldn't find any so I decided to go with the brand colors generally creating my own Code in the meantime. I also used the 8px grids. attach above is also the link to the file uploaded.